[Warhammer]
** Magdar Updated to VC100

++ Added .create command to AccountCacher (.create 'username' 'password')
** New Loots Loading,L oading loot faster

+++ WarhammerV2/Trunk/WorldServer/NetWork/Handler/CharacterHandlers.cs
@@ -220,7 +220,7 @@ namespace WorldServer
         {
             GameClient cclient = client as GameClient;
 
-            cclient.State = eClientState.CharScreen;
+            cclient.State = (int)eClientState.CharScreen;
 
             UInt16 Operation = packet.GetUint16();
